✨ 데이터 추가 신청, 신청 목록 확인, 신청 승인 기능 개발
.

+++ ezdg-auto/src/main/java/com/openmind/ezdg/user/repository/DataApplyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DataApplyRepository extends JpaRepository<DataApplyEntity, Integer> {
+public interface DataApplyRepository extends JpaRepository<DataApplyEntity, Long> {
 }
